refactor(frontend): extract shared tab manager factory
Eliminate TypeScript duplication between modalFact and modalPlan by
creating shared tab manager factory function.

Changes:
- Create shared/tabManagerFactory.ts with createTabManager()
- Refactor modalFact/tabManager.ts to use factory (163 → 30 lines)
- Refactor modalPlan/tabManager.ts to use factory (165 → 34 lines)
- Bundle size reduction: 44.12 KB → 42.68 KB gzip (-1.44 KB)

Resolves known issue #3 from PR #352 (TypeScript duplication).
